requirements.txt
# Django Core
Django==5.2.5
asgiref==3.9.1
sqlparse==0.5.3

# Channels (WebSockets)
channels==4.3.1
daphne==4.2.1
channels-redis==4.3.0
twisted==25.5.0
autobahn==24.4.2
txaio==25.6.1

# Database
psycopg2-binary==2.9.10
dj-database-url==3.0.1

# Security & Encryption
cryptography==45.0.6
pyOpenSSL==25.1.0
cffi==1.17.1
pycparser==2.22

# Configuration
python-decouple==3.8

# Redis
redis==6.4.0
msgpack==1.1.1

# Utilities
attrs==25.3.0
hyperlink==21.0.0
idna==3.10
service-identity==24.2.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
typing-extensions==4.14.1
tzdata==2025.2

# Legacy (pueden ser removidos si no se usan)
automat==25.4.16
constantly==23.10.4
incremental==24.7.2
zope.interface==7.2

# Production (agregar estos)
gunicorn==21.2.0
whitenoise==6.6.0